Move most and least accurate; make much shorter
